feat: restructured and cleaned source folder
