+++ global/apps/authup/src/apps/client-web/index.ts
@@ -1,9 +1,10 @@
 /*
- * Copyright (c) 2023.
+ * Copyright (c) 2022.
  * Author Peter Placzek (tada5hi)
  * For the full copyright and license information,
  * view the LICENSE file that was distributed with this source code.
  */
 
-export * from './execute';
-export * from './types';
+export * from './config';
+export * from './command';
+export * from './constants';
