+++ global/apps/authup/src/apps/constants.ts
@@ -1,12 +1,11 @@
 /*
- * Copyright (c) 2022.
+ * Copyright (c) 2023.
  * Author Peter Placzek (tada5hi)
  * For the full copyright and license information,
  * view the LICENSE file that was distributed with this source code.
  */
 
-export * from './module';
-export * from './reset';
-export * from './start';
-export * from './type';
-export * from './utils';
+export enum AppPackageName {
+    CLIENT_WEB = '@authup/client-web-app',
+    SERVER_CORE = '@authup/server-core-app',
+}
