+++ global/apps/authup/src/apps/index.ts
@@ -5,6 +5,6 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-export * from './config';
-export * from './commands';
+export * from './client-web';
+export * from './server-core';
 export * from './constants';
