+++ global/apps/authup/src/apps/server-core/config.ts
