--- global/apps/authup/src/constants.ts
@@ -1,16 +1,0 @@
-/*
- * Copyright (c) 2023.
- * Author Peter Placzek (tada5hi)
- * For the full copyright and license information,
- * view the LICENSE file that was distributed with this source code.
- */
-
-export enum AppPackageName {
-    CLIENT_WEB = '@authup/client-web-app',
-    SERVER_CORE = '@authup/server-core-app',
-}
-
-export enum AppID {
-    SERVER_CORE = 'api',
-    CLIENT_WEB = 'ui',
-}
