--- global/apps/authup/src/packages/api/commands/reset.ts
@@ -1,23 +1,0 @@
-/*
- * Copyright (c) 2022.
- * Author Peter Placzek (tada5hi)
- * For the full copyright and license information,
- * view the LICENSE file that was distributed with this source code.
- */
-
-import type { ChildProcess } from 'node:child_process';
-import consola from 'consola';
-import type { CommandExecutionContext } from '../../../command';
-import { ServerCommand } from '../constants';
-import { executeServerCommand } from './module';
-import { logChildProcessOutput } from './utils';
-
-export async function resetServer(ctx?: CommandExecutionContext) : Promise<ChildProcess> {
-    consola.info('Server: Cleanup configuration, database, ...');
-    const childProcess = await executeServerCommand(ServerCommand.RESET, ctx);
-    consola.success('Server: Cleaned up');
-
-    logChildProcessOutput(childProcess);
-
-    return childProcess;
-}
