--- global/apps/authup/src/packages/api/commands/start.ts
@@ -1,22 +1,0 @@
-/*
- * Copyright (c) 2022.
- * Author Peter Placzek (tada5hi)
- * For the full copyright and license information,
- * view the LICENSE file that was distributed with this source code.
- */
-
-import consola from 'consola';
-import { ServerCommand } from '../constants';
-import { executeServerCommand } from './module';
-import type { ApiStartCommandContext } from './type';
-import { logChildProcessOutput } from './utils';
-
-export async function startServer(ctx: ApiStartCommandContext) {
-    consola.info('Server: Starting...');
-    consola.info(`Server: Port ${ctx.env.PORT}`);
-
-    const childProcess = await executeServerCommand(ServerCommand.START, ctx);
-    consola.success('Server: Started');
-
-    logChildProcessOutput(childProcess);
-}
