--- global/apps/authup/src/packages/api/commands/type.ts
@@ -1,13 +1,0 @@
-/*
- * Copyright (c) 2023.
- * Author Peter Placzek (tada5hi)
- * For the full copyright and license information,
- * view the LICENSE file that was distributed with this source code.
- */
-
-import type { CommandExecutionContext } from '../../../command';
-
-export type ApiStartCommandContext = CommandExecutionContext<{
-    PORT: number,
-    WRITABLE_DIRECTORY_PATH: string
-}>;
