--- global/apps/authup/src/packages/ui/commands/type.ts
@@ -1,19 +1,0 @@
-/*
- * Copyright (c) 2023.
- * Author Peter Placzek (tada5hi)
- * For the full copyright and license information,
- * view the LICENSE file that was distributed with this source code.
- */
-
-import type { CommandExecutionContext } from '../../../command';
-
-export type UIStartCommandContext = CommandExecutionContext<{
-
-    PORT: number,
-
-    HOST: string,
-
-    API_URL: string,
-
-    PUBLIC_URL: string
-}>;
