--- global/apps/authup/src/packages/ui/constants.ts
@@ -1,10 +1,0 @@
-/*
- * Copyright (c) 2022.
- * Author Peter Placzek (tada5hi)
- * For the full copyright and license information,
- * view the LICENSE file that was distributed with this source code.
- */
-
-export enum UICommand {
-    START = 'start',
-}
