+++ global/apps/authup/src/utils/index.ts
@@ -5,6 +5,7 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
+export * from './execute';
 export * from './line-breaks';
 export * from './process-output';
 export * from './stringify-object-args';
