+++ global/packages/core/src/utils/app-id.ts
@@ -1,0 +1,13 @@
+export function parseAppID(id: string) {
+    const match = id.match(/([^:/]+)[:/]([^:d/]+)/);
+    if (!match) {
+        return undefined;
+    }
+
+    const [, group, name] = match;
+
+    return {
+        group,
+        name,
+    };
+}
